requirements.txt
alabaster==0.7.16
antlr4-python3-runtime==4.11.0
babel==2.16.0
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
docutils==0.21.2
flask==3.1.0
flask-sqlalchemy==3.1.1
greenlet==3.1.1
idna==3.10
imagesize==1.4.1
importlib-metadata==8.5.0
itsdangerous==2.2.0
jinja2==3.1.4
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.2.1
packaging==24.2
psycopg2==2.9.10
pygments==2.18.0
python-dotenv==1.0.1
requests==2.32.3
snowballstemmer==2.2.0
sphinx==7.4.7
sphinx-autodoc-typehints==2.3.0
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
SQLAlchemy==2.0.36
sympy==1.13.3
tomli==2.2.1
typing-extensions==4.12.2
urllib3==2.2.3
werkzeug==3.1.3
zipp==3.21.0
